requirements.txt
fastapi
langchain
langchain-community
langchain_astradb
langchain_core
python-dotenv
pandas


-e . 